Remove go 1.11 price functions (#1716)
We are only supporting Go 1.12+ so it is no longer necessary to have price/mul_legacy.go
